A dramatic change to the structure of GEMsolve. Model types (previously run types) and runs (tmg, reo, dis) are now gone. In their place we have 'experiments' and 'steps'.
An 'experiment' is a complete run through of timing-reoptimisation-dispatch. There are 'steps' within each experiment - the steps are 'timing', 'reopt' and 'dispatch'. There are outcomes, which are individual scenarios in the stochastic sense, and there are 'outcomeSets' which are sets of outcomes. So, you bunch together outcomes into a set to run a stochastic model.

You define a set of experiments, and for each step of each experiment you define the outcomeSets you'd like to use.

This is a likely to be refined in coming days so stay tuned!