A bunch of minor tidying up - especially in GEMstochastic, GEMdeclarations and GEMsolve. A non-exhaustive list includes: - removed hydroYrForDispatch - changed setOutcomes to mapOutcomes - changed set_outcomeWeight to weightOutcomesBySet - changed allRuns to allSolves - changed sumRuns to sumSolves - changed numRuns to numSolves
Outcome weights are pro-rated (in GEMsolve) so that the weights sum to exactly one for each element if outcomeSets.

The flag for slacks has been split in 2 - one flag for slacks and one for penalties.

A new parameter called allModelledHydroOutput has been created to capture the actual hydro data used in each and every solve. This gets dumped to a GDX file and can be used to check the code is doing what you think you told it to do.

Note that the elements of set hY called average and multiple no longer exist - they've been gone for a while now.