Edited GEMstochastic and GEMdata to include a little more reporting of input data at end of GEMdata.gms.
More specifically, a file called 'Stochastic summary.txt' is now created in the 'Input data checks' folder to report how the experiment-outcomeSet-outcome and outcome weights/factors etc are mapped and related.

Also, the user is required to specify a single outcome in GEMstochastic (this will eventually come from Oasis) for the purpose of reporting summaries of energy and peak load - see reportingOutcome(outcomes).