Removed 5 fundamental sets from the input GDX file as their membership never changes - y, ild, col, m and geo. All but y is hard-coded in GEMdeclarations. Membership of set y is established by user and specified in GEMsettings.
Also separated the energy demand (load) from the main input GDX file, i.e. th euser now needs to specify 2 input GDX files - GEMinputGDX and GEMdemandGDX. The reason for this is that demand data is typically specified according to one of several development/growth assumptions. It's simpler to read in the desired assumed demand growth rather than read in a 'standard' growth profile and then read in another as an override.