Removed several technologies upon which no plant were defined: bioPkr, coalCog, WindDG, IGCC, IGCC_Lig and HydDG. Removed several technologies and the plant defined on them: - coal_DY and the 4 x 245 MW generating units defined on coal_DY, - IGCC_CCS and the 4 x 400 MW generating units defined on IGCC_CCS, - IGCC_LigCCS and the 1 x 400 MW generating units defined on IGCC_LigCCS, - CCGT_CCS and the 2 x 410 MW generating units defined on CCGT_CCS, - Wave and the 4 x 38 MW defined on wave, and - Solar and the 6 x 50 MW plant defined on solar.
Removed all sets and parameters associated with the CCS technology - CCStech(k), i_CSSfactor(k), i_CCScost(k) and CO2captureAndStorageCost(g,y).

Removed the 360 MW plant called TCCseas.

Deleted the fuels called BioD and solar; and changed the name of bio to other othCog, changed the name of oth to othCog, and changed the name of mrn (marine) to wave.

There are now 19 technologies and 10 fuels. There are now 256 generating plant, of which 45 already exist.