Got rid of i_minUtilByTech(y,k) and minUtilTechs(g). Replaced i_minUtilisation(g) with i_minUtilisation(g,y). And made corresponding changes to various codes.
Upshot of this is that the minutilisation constraint requires data to be provided by plant and year only - nothing else.

Also reduced value of min utilisation to 0.2 for coal, lignite, and CCGT, and to .05 for OCGT.

Added a "MW built by region and tech as a % of MW available to be built" table to the summary report file.