GEMreports now produces a second csv file called "Results to be plotted - XXX.csv" which can be found in the 'Processed files' directory. It also produces a batch file called GEMplots.bat that gets placed in the archive directory.
Provided GEMplots.exe (not yet in repository but will soon be) is located in the GEM matlab codes folder, the aforementioned batch file can be double-clicked and a bunch of plots will be generated. GEMplots.exe requires the MCR to be installed.

The model type for GEM and DISP, i.e. MIP or RMIP, is stamped into the solve summary report.