Modified the IL plant data and made minor tweaks to the comments throughout the 4 main programs - GEMdeclarations, GEMdata, GEMsolve, and GEMreports. Also changed the alias of hY from hY1 to hYY.
IL plant: Previously there was 1 existing IL plant (IL0) of 160 MW located at Otahuhu and 3 new ones (IL1, IL2, and IL3) each of 50 MW with IL1 and IL2 located at Otahuhu and IL3 at Islington. The earliest commissioning year was, respectively, 2016, 2021 and 2026. The capital cost ($/kW) was, respectively, 400, 440 and 484.

There are now 18 new IL plant - 1 for each of the regions in the 18-region version of GEM. The earliest commissioning year of each is set to 2012, the capital cost of each is set to $440/kW, and the nameplate capacity is set to 5% of the 2012 load (GWh) in each region divided by 8.6. More specifically, the short name of new IL plant and its nameplate MW are: ILnthd 10, ILnshr 8, ILakld 51, ILwaik 21, ILbopy 18, ILhbay 10, ILtari 5, ILbtpe 12, ILwgtn 19, ILnmlb 7, ILkiki 1, ILwcst 2, ILchch 28, ILwtki 5, ILotag 10, ILslnd 6, ILflnd 0, and ILtiwa 32.