Modified the way in which GEMreports is passed configuration info: - added a new .inc file - GEMreportSettings.inc - GEMdata produces a new file - "Configuration info for GEMreports - XXX.inc", which resides in the 'Input data checks' folder. - GEMreports no longer includes GEMsettings.inc, GEMpathsAndFiles.inc, or GEMstochastic.inc.
The motivation for these changes is to decouple the invocation of GEMreports from that of invoking GEMdata/GEMsolve. The new file called "Configuration info for GEMreports - XXX.inc" contains the information from GEMstochastic.inc and GEMsettings.inc that is required by GEMreports.gms. GEMreportSettings.inc contains info that is specific to a given run of GEMreports and is independent of, and possibly unknown, at the time GEM is solved.

All of this is being done in anticipation of GEM being operated from within emi.