Modified GEM to keep track of all run versions associated with a GEM run:  - a file called runVersions.txt is now created, which will contain a list of all run versions in a run, as well as a long description of each version and the user-assigned RGB codes.  - two more $setglobals have been added to GEMpathsAndFiles.inc to follow $setglobal runVersionName mds5:    - $setglobal runVersionDesc  "High gas discovery"    - $setglobal runVersionRGB   "255.0.255"
The motivation for the above is to simplify and automate the process of setting up GEM to produce reports for all user-specified solutions drawn from previously solved runs and run versions.

The archiving of the solver options file has been ceased as these are now created on the fly, based on statements in GEMsolve, which itself is archived.